// types.ts
export interface DistrictStats {
  population: string;
  area: string;
  mahallas: number;
  economyType: string;
}

export interface Official {
  name: string;
  position: string;
  image: string;
}

export interface Landmark {
  name: string;
  description: string;
  image: string;
}

export interface FamousPerson {
  name: string;
  contribution: string;
  image: string;
}